js: Split out muted_users_ui from muting_ui module.
